refactor:update md_spectra
